Extracts ExcludedResource
